+++ qt/widgets/common/CMakeLists.txt
@@ -9,7 +9,6 @@ set(
   src/AlgorithmSelectorWidget.cpp
   src/AlternateCSPythonLexer.cpp
   src/ConvolutionFunctionModel.cpp
-  src/BackendQt.cpp
   src/BasicFitOptionsBrowser.cpp
   src/BatchAlgorithmRunner.cpp
   src/BoolPropertyWidget.cpp
@@ -351,7 +350,6 @@ set(
   src/AlgorithmPropertiesWidget.cpp
   src/AlgorithmRunner.cpp
   src/AlternateCSPythonLexer.cpp
-  src/BackendQt.cpp
   src/BatchAlgorithmRunner.cpp
   src/BoolPropertyWidget.cpp
   src/DropEventHelper.cpp
