+++ qt/widgets/mplcpp/CMakeLists.txt
@@ -3,6 +3,7 @@ set(
   LIB_SRCS
   src/Artist.cpp
   src/Axes.cpp
+  src/BackendQt.cpp
   src/Colors.cpp
   src/ColorbarWidget.cpp
   src/Colormap.cpp
